refactor:adust folders and config path
