+++ babel.config.js
@@ -9,7 +9,7 @@ module.exports = function (api) {
                 {
                     alias: {
                         '@components': './src/common/components',
-                        '@routes': './src/infra/config/routes',
+                        '@router': './src/router',
                         '@types': './src/common/types',
                         '@models': './src/common/models',
                         '@utils': './src/common/utils',
