+++ tsconfig.json
@@ -5,7 +5,7 @@
         "baseUrl": "./",
         "paths": {
             "@components/*": ["src/common/components/*"],
-            "@routes/*": ["./src/infra/config/routes/*"],
+            "@router/*": ["./src/router/*"],
             "@types/*": ["src/common/types/*"],
             "@models/*": ["src/common/models/*"],
             "@utils/*": ["src/common/utils/*"],
